// app/features/events/model/event.ts
import type { PaginatedResponse } from '~/shared/model/pagination'
import type { AuthUser } from '~/features/auth/model/user'

export const EventGameType = {
  QUIZ: 'quiz',
  POOL: 'pool',
  MAFIA: 'mafia',
  POKER: 'poker',
  DARTS: 'darts',
} as const

export type EventGameType = (typeof EventGameType)[keyof typeof EventGameType]

export const EventGameStatus = {
  DRAFT: 'draft',
  PUBLISHED: 'published',
  CANCELLED: 'cancelled',
  COMPLETED: 'completed',
} as const

export type EventGameStatus = (typeof EventGameStatus)[keyof typeof EventGameStatus]

export type RegistrationStatus =
  | 'ACTIVE'
  | 'PARTICIPANT'
  | 'WAITING'
  | 'WAITLIST'
  | 'WAITING_LIST'
  | string

export type Event = {
  id: string
  imageUrl: string | null
  imageHash: string
  title: string
  city: string
  address: string
  gameType: EventGameType
  gameRules: string
  features: string
  startsAt: string
  endsAt: string
  location: string
  participantLimit: number
  initialDepositAmount: number
  maxParticipants?: number
  seatsPerTable?: number
  tableCount?: number
  pointsForParticipation: number
  status: EventGameStatus
  createdAt: string
  updatedAt: string
  reminderSentNow?: boolean
  reminderSent24h?: boolean
  reminderSent2h?: boolean
  reminderSent15m?: boolean
  _count: {
    registrations: number
  }
}

export type EventsResponse = PaginatedResponse<Event>
export type EventResponse = Event
export type TemplatesResponse = PaginatedResponse<Event>

export type GetEventsParams = {
  gameType?: EventGameType | string
  status?: EventGameStatus | string
}

export type UpdateEventParams = { id: Event['id'] }

export type UpdateEventPayload = {
  title: Event['title']
  city: Event['city']
  gameRules: Event['gameRules']
  features: Event['features']
  address: Event['address']
  startsAt: Event['startsAt']
  gameType: Event['gameType']
  participantLimit: Event['participantLimit']
  initialDepositAmount: Event['initialDepositAmount']
  imageUrl: Event['imageUrl']
  imageHash: Event['imageHash']
}

export type CreateEventPayload = {
  title: Event['title']
  city: Event['city']
  gameRules: Event['gameRules']
  features: Event['features']
  address: Event['address']
  startsAt: Event['startsAt']
  gameType: Event['gameType']
  participantLimit: Event['participantLimit']
  initialDepositAmount: Event['initialDepositAmount']
  imageUrl: Event['imageUrl']
  imageHash: Event['imageHash']
  seatsPerTable: Event['seatsPerTable']
  isTemplate?: boolean
}

export type ReorderParticipantsPayload = {
  participants: {
    userId: string
    position: number
  }[]
}

export type FinalizeEventResponse = {
  success: boolean
}

export type GetEventParticipantsResponse = {
  event: Event
  participants: Participant[]
}

export type Participant = {
  id: string
  userId: string
  eventId: string
  status: RegistrationStatus
  createdAt: string
  cancelledAt: string
  position: number | null
  user: AuthUser
}